created controllers
- created restaurant and dish controllers for admin
- changed delete method in DataJpaDishRepository
- refactored RestaurantRestController
- fixed update method in AbstractUserController

+++ src/main/java/ru/iliketobreathe/restaurantvote/repository/dish/DataJpaDishRepository.java
@@ -32,12 +32,11 @@ public class DataJpaDishRepository {
                 .orElse(null);
     }
 
-    public boolean delete(int id) {
-        return crudRepository.delete(id) != 0;
-    }
-
-    public Dish get(int id) {
-        return crudRepository.findById(id).orElse(null);
+    public boolean delete(int id, int restaurantId) {
+        if(get(id, restaurantId) != null) {
+            return crudRepository.delete(id) != 0;
+        }
+        return false;
     }
 
     public List<Dish> getAll(int restaurantId, LocalDate date) {
